requirements.txt
opencv-python
numpy
pillow
ultralytics
beautifulsoup4
requests
selenium
googlesearch-python
groq